ADHOC fix (Snapshots-Tagging): Fix tagging snapshots by constructing command with eval.
The aws api allows pushing tags to snapshots.
The --tags parameter requires a list of tags in the form
Key=TagKey,Value=TagValue.
In case of multiple tags the api takes the Key of the first tag
and the Value of the last tag and pushes only on tag. This is fixed by
constructing the command with eval

As part of the fix the "=" needs to be removed as its not part of the
syntax defined for the aws api